[lucet-runtime-internals] terminate the stack correctly
This places the pointers to the parent and child contexts _above_ the two zero words used to
terminate the call stack. This is the location where memory arguments are expected, which isn't
completely relevant since `lucet_context_backstop` is written in assembly, but it at least prevents
gdb and other unwinding tools from mistaking the pointers for return addresses.