[lucet-runtime] implement counter for unwinding nested hostcalls
This doesn't work yet, because the prologue for guest functions is trying to push a frame
pointer. At least in the new test added in this commit, the value of `rbp` at entry to the callback
`guest_func` is zero, which screws up how the debugger (and presumably the unwinding runtime)
interpret the stack.